fix read me and cli flag
